Only email notify on consecutive bad health issues
We used to email on any issue, even if it might have been just
temporary. This is uneccessary and creates false concerns. So now
we're only performing email notification on consecutive issues.

This has involved modifying how we check for issues i.e. we return
the string of the issue rather than a bool. We also provide this
to the badHealthCallback, where we can handle with health
notification if we wish.