Save broadcast issues tally
It turns out that we weren't getting any badHealthEvents published
and subsequently never transitioning to the Unhealthy states
because we weren't actually saving the changes to the issues tally
to the datastore. This change fixes this.

It has required that we alter the interface of HandleHealth
slightly to require a parameter of Store interface type.