Fix direct state behaviour
The behaviour has been incorrect. To start, we should not stop
the broadcast on exit of the directLive state. Maybe once upon
a time this worked because there was no state other than idle
that we'd exit to, but now we have a directLiveUnhealthy. So, we
only try to stop the broadcast (if exists) on entry to directIdle.

Secondly, the directLiveUnhealthy similarly should not try to stop
the broadcast on exit. It should also implement the fixable state
so that it can try to fix the hardware in this state.