Site Menu: Dispatch Events on site changes (#88)
When the site is changed, the sitemenu will now dispatch a site-change event instead of forcing a refresh. This prevents unwanted form resubmissions on pages, and gives developers more control over how to handle a site change. This event also carries a previousSite detail, allowing a page to revert back to a previous site if a site change was not wanted.