Site Menu: Default to refreshes (#88)
To make implementation of the site menu easier, it will default to just refreshing the page. However, if the page requires a more careful approach, the site menu can be set to have a custom handling, in which case it fires events. This means that pages do not have to be changed to allow for this new upgrade to be made.