Broadcast: Implement custom site-change handling (#88)
The broadcast page will now listen for the site-change event and handle appropriately. The page will warn the user that unsaved data might be lost, and give them the option of changing back to the previous site.
If the user confirms a site change, the current window location is resent to /admin/broadcast (not using a refresh), which will not resubmit any data.