Add token generation button to broadcast page
To associate and authenticate a broadcast with a particular
YouTube account, we need to generate a token for the account.

We're providing a button on the broadcast page that performs the
generation and stores the account email in the broadcast config,
which can be used to later retreive the token for refresh and/or
use.