Check controller state before trying to start broadcast
We're now checking the state of the controller before we try to
start a broadcast. If the we're not in a good state we publish
a new event "controllerFailureEvent".

The broadcast state machine now handles the controller failure
event, and transitions the broadcast into the idle or slate state.
This handler also calls onFailureClosure to disable the broadcast
on the first fail.

In this case we want to fail straight away, so the onFailureClosure
has been adapted to take a boolean to indicate we don't want
retries.