Improve permanent live unhealthy state behaviour
We're now allowing only limited health fixes when in the permanent
live unhealthy state. Fixes include hardware reset and requesting
the forwarding service for stream (in case problems are on this
front).

In the case that we max out our fixes allowance, we
transition to a new permanent failure state. In this state we
request that the forwarder goes to slate mode. We've also
introduced a state to the broadcast object "InFailure" which is
true if the broadcast is in a failure state. This is accessible
from the broadcast UI. We'll need to uncheck this once we've
resolved and issues and would like to resume normal behaviour.