Bench: Prevent Changing of Youtube Channel (#155)
When the broadcast save action is called we will get the previously associated account for the broadcast and update the current config with that value to prevent users from changing this. The value can only be changed when the generate-token action is called.
We will also not parse the value from the form to ensure this value can "never" be leaked through.