Slate on hardware failure for permanent live states
We're now handling hardware failure events when we're in the
permanent live/live unhealthy and slate to live states. Under
these circumstances, we should stay in slate. We can do this by
entering the vidforwardPermanentInFailure state in which case we
wait until admin has taken the broadcast out of this state i.e.
maybe fixed the problem.