Define Notifier interface
It's useful to have a Notifier interface. Firstly, it allows us
to swap out the real implementation with a mock implementation for
testing purposes. Secondly, it's possible down the track that we'll
want to have different notifier implementations, maybe for example
an SMS implementation.

Therefore, we have re-labelled our current notifier as the
MailJetNotifier. Also, we have removed the Init method and instead
opted for implementation specific initialisation i.e. have a
NewMailJetNotifier constructor accept the functional options. This
is required given that options must be specific to a particular
implementation.