Improve testing facilities
This change aims to improve aspects of the broadcast testing
facilities.

To start, we're adding logOutput and notifier fields to the
broadcastContext. This will allow us to specify different log and
notification implementations for the context. For example t.Log and
a mock notifier implementation when testing.

We're also now providing some factories for common broadcast context
arrangements.

We've added mock eventBus and notifier implementations, and a log
output implementation, logRecorder, which will allow us to validate
logging.

Some other smaller changes:
* unexporting the dummyManager
* modifying dummyManager.Save method to better reflect our real
implementation.
* providing constructors for the dummyService and dummyStore
implementations