Add broadcastSystem abstraction
This change creates an abstraction of the broadcast system. Up
until this point, we've had to manually "construct" the broadcast
system with all of the components and subscribe event handlers etc.
This was happening for the production check handler code, as well
as all the testing code.

The intention here is that we can simplify this process, and
encourage more use of this in integration type testing.

Also, health, status and chat checks are now part of the broadcast
state machine; there's no reason why these had to be separate.