Start adding configuration sanitisation
We're starting with checking the hardware is defined for a given
broadcast, otherwise we disable. We're doing this with an
invalidConfigurationEvent. This can be used for various
configuration issues, and for convenience also impelements the
error interface.

We're adding a basic test case that highlights the ease of
integration type testing using the new broadcastSystem abstraction
added in the prior change.