Provide  NewConfigurableRecoveryHandler
This provides a nice interface for creating a recovery handler. It
works on the assumption that there's some common things we might
like to do for panic recover i.e. first decide whether we should
handle the particular panic, format a message in response to the
panic, log that message, notify through email/sms of that message,
produce a HTTP status and finally maybe ask for help from further,
more custom handlers.

We provide options to customize each of these steps using the
functional options pattern.

We're now using this handler factory to set up a handler for
oceantv. We've added an additional handler that is invoked for
errNoGlobalNotifier type panics, in which case we initialise
the global notifier in attempt to remediate the problem.