tv: only create a new broadcast when one doesn't already exist (#271)
* tv: only create broadcast when needed

this was done to prevent duplicate broadcasts

* tv: remove CreateBroadcastIfNeeded, do check in CreateBroadcast

* tv: allow creation of new broadcast if status is empty

This was done because if we don't have status information, we can't proceed with an existing broadcast

* tv: modify log in test to pass

A small update was needed for a test to pass

* tv: wrap error in getStreamStatus

this was done because the error could be useful

* tv: create new broadcast if SID is empty

This was done because the broadcast config would be reused but the stream ID could be empty for some reason. When this happens we can create a new broadcast and stream.

* tv: simplify code

This was done to address PR feedback about simplicity.

* tv: invert broadcast exist check to simplify code

Done to address PR feedback, making code neater.

* tv: put broadcast reuse check in a function, add tests

This was done to simplify the code and increase test coverage.

* tv: check that we don't reuse an old broadcast

This was done so that we don't reuse old broadcasts if they exist. Only use broadcasts from today.
A TimeCreated field was added to broadcast config to record when the broadcast has been created, this is what is checked.
Tests were also added.

* tv: add GetSchduledStart to BroadcastService

This was done so that we could ue the broadcast object's scheduled start time instead of creating an extra field in the BroadcastConfig.

* tv: update tests to use functional options

This was done so that I didn't have to add another dummy service to test good statuses with.

* tv: log error when reusing an old broadcast, bump version

* tv: add integration test for CreateBroadcast

This was done to test the CreateBroadcast function within the context of a broadcast system.

* tv: return error for multiple broadcasts with the same ID

This was done to alert us of any strange behaviour as this shouldn't be possible.