Add slate options
This change modifies the ForwardingService interface's Slate method
to accept SlateOptions. We've added an option WithType that will allow
us to specify the type of slate. This has been motivated by the fact that
we'll have a permanent broadcast low voltage state and may wish to
indicate this by a particular slate.