Add hardware voltage logic to broadcasting
Up until this point the broadcasting logic has had no awareness
of hardware voltage. Therefore failures to broadcast, or issues
during broadcasting that were due to low voltage, were not handled
gracefully.

This change attempts to improve matters by checking the hardware
voltage when starting a broadcast, or in the case of permanent
broadcasts, when transitioning from slate to live.

In the event that the check indicates we have low voltage, we
enter a hardware voltage recovery state. In the case of
permanent broadcasting, we can also indicate a low voltage state
through setting a low voltage slate to display.

During recovery, timeEvents invoke voltage checks. Once we reach
the required voltage we assume normal function. If a timeout is
exceeded, we enter a failure state.

Some more functionality and testing could be added in regards to
controller or charging faults for permanent broadcasting. For
example, slate modes could be added for these conditions as well.