Embed common state properties (#330)
* Extend hardwareManager interface

This change extends the hardwareManager interface by adding voltage,
alarmVoltage and isUp methods. These additions have been motivated
by voltage handling in the broadcast functionality.

* Add slate options

This change modifies the ForwardingService interface's Slate method
to accept SlateOptions. We've added an option WithType that will allow
us to specify the type of slate. This has been motivated by the fact that
we'll have a permanent broadcast low voltage state and may wish to
indicate this by a particular slate.

* Embed common state properties

This change attempts to simplify states and reduce boilerplate by
embedding common fields into states. This allows the implicit
inclusion of enter and exit methods, as well as fields and methods
require for states that timeout.

---------

Co-authored-by: Trek H <trek.hopton@gmail.com>