ausoceantv: add super basic home and watch pages (#357)
* AusOceanTV: Extend LitElement with Tailwind styles.

This change adds support for tailwind styling in lit elements using
the Vite Dev tools and PostCSS.

* AusOceanTV: make style an optional argument for TailwindElement

This change means that to create a TailwindElement, you no longer require an external style to be passed, which we will not always need.

* ausoceantv: style temporary main page with tailwind

This was done to link to the streams list and watch pages.

The custom web component logo-holder is not used but I'm leaving it there to be a template for future elements.

* ausoceantv: link to home page

Improve style of welcome page, link to home page.
The home page will be where logged in users start. It will primarily provice a list of streams.

* ausoceantv: basic movement between main pages

Added the youtube iframe player on the watch page.

* ausoceantv: add basic watch page, remove empty files

This was done because I forgot to add the watch.html file to the commit.
Also, we don't need the empty lit element. We can just remember that they go in src with their own dir to contain the ts and css.

---------

Co-authored-by: David Sutton <david@ausocean.org>
Co-authored-by: David (AusOcean) <davidsutton@ausocean.org>