Go to failure state instead of disabling broadcast
Up until this point, critical failures i.e. those that cause repeated
broadcast failure, have resulted in disabling of the broadcast.
Unfortunately, this meant that any stored events at the time of
disabling would not get processed, such as hardware shutdown requests.
This is because the state machines of disabled broadcasts do not spin up.

Furthermore, there should probably be a differentiation between a
broadcast that is in a critical failure state, as opposed to one that is
just inactive.

For the aforementioned reasons, this commit changes the behaviour of the
broadcast state machine when it struggles to start a broadcast. If we
exceed the allocated number of attempts, then a critical failure event
is emitted. In response to this event, the broadcasts will move to a
failure state, which is indicated in the broadcast UI via a tick box.
This must then be unchecked by admin to put the broadcast into a
startable state.

Direct broadcasts did not have a failure state, so this has been added.