Register events using to dedicated registry
The current pattern of adding an event type and then adding to the
stringToEvent function map was error prone and clunky. By using
a registry we can register events immediately, which is much more
intuitive and clear. The stringToEvent function then just uses the
registry to retrieve an event of the provided type.

The registry employs a singleton i.e. the registry is only ever
instantiated once, and cannot be externally re-instantiated or
replaced.