Register events and states using registry
The current pattern of adding an event or state type and then adding to
the stringTo function map was error prone and clunky. By using a
registry we can register events immediately, which is much more
intuitive and clear. The stringTo functions then just uses the registry
to retrieve an event of the provided type.

The registry employs a singleton i.e. the registry is only ever
instantiated once, and cannot be externally re-instantiated or
replaced.