Add deployment script for cloud projects
This change adds a deploy.sh script which can be called from a cmd/ directory to build and
deploy a service. This script can deploy to either a development mode (dev-dot-...) or to
the most recent version (scraped from main.go version) with the version being promoted.

This script has been added to npm run deploy and deploy:dev for both ausoceantv and oceanbench.