oceantv: split stopping state into shutdown and power off
We would like to start doing a proper shutdown to power down the cameras
as opposed to just cutting the power. To do this we have decided to have two
substates in the stopping state i.e. shutdown and power off.

Another change that has resulted is that the restart state actually has
substates i.e. stopping and starting.

These changes have resulted in a cascade of changes especially to testing.