Add device throughput graphing
This is useful to see throughput characteristics over time.

We're doing this by adding a quasi-pin i.e. it's not a real sensor pin, but
we can set the pin to "throughput" on the UI for the device, and then
have special handling of this case to use the throughputs endpoint
which gets historical data to provide for graphing. The process is very
similar to the data endpoint handling however, the datastore ops are
different to get the data.