bench: keep track of livestream visits with URL (#402)
* bench: keep track of livestream visits with URL

This was done because it should give us insights into how many people are visiting our streams through the URL redirect. The more user analytics the better.

* bench: remove unused writeHTTPErrorAndLog function

This was done to clean up dead code.

* model: remove error from update function in incrementVisitCount

* model: simplify PutVariableInTransaction

The update function is now directly used in store.Update.

* bench: log if parsing visit count fails

This was done so we don't miss the error.

* bench: improve flow for PutVariableInTransaction

This was done to limit calls to the datastore based on PR feedback.

* bench: update PutVariableInTransaction comment

This was done as per PR feedback

* bench: minor version bump

This is for link counter functionality for redirected live links.