AOTV: Update stripe subscription to cancel (#420)
This change update the subscription on the stripe server to make sure that we are in sync when cancelling
to avoid users getting charged when it seemed like they had cancelled the subscription.

This change also adds a few checks and conditional rendering in the profile-details element to make it easier
to see the state of the subscription.