ausoceantv: user survey (#470)
* ausoceantv: add survey page

This was done to capture user demographic data. We still need to link to it and handle submissions.

* ausoceantv: add survey page go handlers

If the user is over a day old, it will be redirected to the survey page.
Upon submission of the survey, the interest and location will be saved in the demographic data.

* ausoceantv: always return JSON for survey check handler

This was done so the client side code always has valid JSON as a response. This stops unnecessary errors in the browser.