ausoceantv: use google maps API for location field autocomplete in survey (#475)
* ausoceantv: add autofill to location field for survey

This was done to make it easier for the user to enter their location.

* ausoceantv: capture postcode and city/suburb

This was done to increase the granularity of the location for Australian users.

* ausoceantv: add (optional) for user location  field

This was done so that users aren't deterred by having to put in location data which some people may view as too personal.

* ausoceantv: don't only restrict post codes to Australia.

This was done because we could have users from other countries.

* ausoceantv: add back au restriction for survey location

This was done because otherwise it's too annoying to find Australian locations. A following change will be to handle global locations.