ausoceantv: make SubscriberRegion entity, methods and use it in the survey (#490)
* ausoceantv: create SubscriberRegion entity

This was done to store location information for subscribers in a queryable format.

* ausoceantv: add get and put function for SubscriberRegion

This was done to create or update the sub region when the survey is filled out.

* ausoceantv: change references to geocode to region, fix json unmarshalling

This should fix an error where the region fields weren't updating properly. Geocode is not really the right term for what we're doing here, so I've changed it to region.

* ausoceantv: add import for google maps to ts

This was done so that npm run build would work.

* ausoceantv: store user-category as JSON

This was done so that if we want to add more fields in future, we can.

* ausoceantv: rename regions.go to subscriberregion.go

Use Get not GetAll for SubscriberRegion.Get.