otv: handle default substate case for hardware states
There's no normal circumstance where we hit the default case for time event
handling in hardware states that have substates. So we should notify of this
happening and default to something sensible. We can do this be calling the
entry function which will initialise the substate back to the initial state.