ausoceantv: home and watch footer position, scroll free-banner (#497)
* ausoceantv: home and watch footer position, scroll free-banner

The home footer was made to be on screen unless pushed off by content.

The watch footer was made to always be on the bottom.

The free banner was made to scroll with the page while remaining absolutely positioned so that it doesn't affect other elements.

* ausoceantv: adjust padding for breakpoints

Done for aesthetics on smaller devices.