oceantv: extra logs for hardwareStopping sub-state machine (#499)
* oceantv: add logging to hardwareStopping state for debugging

* oceantv: log when timeEvent doesn't fall inside starting time.

This was done because sometimes it's unclear why nothing happens when a time event is handled.

* oceanbench: format broadcast timezone properly

For some reason a timezone of 9.5 was being formatted as +10:30 so I have rewritten the code to be more understandable and hopefully work.

* oceanbench: bump version

* bench: revert timezone from site changes

This was done because it was unecessary since the timezone in set from the client's timezone.