cmd/cloudblue: initial copy from ausoceantv for app structure
Everything that is for AusOceanTV will be stripped, this is just a good starting point for a vite project.

Instead of OceanBench, which has been our cloud UI for a little while, I am experiemnting with a new UI service called CloudBlue (since we call it that anyway),
that will be built in Vite. It may have SSR components but if possible we will keep the SSR light weight.