oceantv: correct secondary broadcast saving
This corrects secondary broadcast saving in a couple of ways. Firstly,
we're no longer using the deprecated saveBroadcast function which
does not use transactions, and secondly, we're correcting the default
save behaviour by create a temporary broadcast manager for the
secondary broadcast so that we don't overwrite the primary. This
is a programming error.