cmd/ausoceantv: if there is no refresh token, prompt consent (#523)
* cmd/ausoceantv: if there is no refresh token, preompt consent

There's no way of getting another refresh token if it's gone for some reason (user clears cookies),
so we need to check and prompt for consent if it's missing to get another refresh token.
Unless we decide we want to store the refresh tokens in our DB.

* cmd/ausoceantv: bump verison

* gauth: only check fro refresh token if session loads without error

This was done to prevent silent errors or unexpected behaviour from nil sessions.