OTV: Don't notify with no shutdown actions
This change stops a notification if the reason that the shutdown failed was due to there being no
shutdown actions provided. This was done so that if cameras are not configured correctly to shutdown that
we won't get notified every time that it turns off if we don't set any shutdown actions.