otv: enable state/event based webhooks
This change provides the option to add event or state based webhooks
(or any other sort of hook for that matter). This is is motivated by the
desire for services such as AusOceanTV and OpenFish to have an
awareness of broadcast events and states.

We have done this by using options to an oceantv service, which then
propagates these hooks to the state machine.

closes #526