otv: improve error event handling
The intention here is to improve the function of events that
represent erroneous conditions and make error information
more available and clearer. We're also trying to use more
consistent patterns around error conditions, and do more
central notification on these error events.

Events representing erroneous conditions are errors themselves.
They also hold errors instead of basic strings. This means we
can chain error events/errors and maintain a history of the
errors that led us to where we are. As such we have unwrapping
functionality to find the initial error event type. We've also added
a Kind() method to the interface that gives us the corresponding notify.Kind so that we can perform a notification of the correct
kind.

Finally we've fixed some fundemental issues around
storing events for processing; we weren't properly storing the
error events internal error information, and would therefore
loose this on the next SM tick.