otv: fix event marshalling/unmarshalling and error matching
Events are complex in that they can be of errorEvent type which
means that they can chain other errorEvents or normal errors.

This change fixes error matching using Is and
marshalling/unmarshalling so that it happen recursively and
reconstruct the events exactly as they were. We've also added
testing to verify things are working as expected.