otv: improve directLiveUnhealthy fix code
In the case that there's a fix failure, we're now publishing a
fixFailureEvent rather than a finishEvent, which implements the
errorEvent interface now. This makes it clear that this is an error
condition and we should respond accordingly, otherwise the SM will
just try to start the broadcast again once it gets back to the
idle state. As part of this we've added a transition to the
directFailure state if we get a fixFailureEvent event in the the
directLiveUnhealthy state.

We've also fixed the attempt count increment logic; this was
incorrect and so we would get less retries than we thought we would.