Bench: Convert more places to UUID (#666)
This change fixes a few bugs, and potential bugs by ensuring more of the
logic was using UUIDs. All of the current datastore broadcasts have been updated
to UUIDs, and so no handling for other versions is guaranteed to work, so local
filestores may have to be updated. Saving a broadcast without a UUID should still
generate a new UUID for it.

closes #666