OTV: Add sub claim to JWT
This change adds a subjecter interface which returns a subject string
to be used as the sub claim in a JWT. This is used by AusOceanTV to
verify that the youtubeID being sent was signed by a client with the
secret key. This prevents an attacker from sending an arbitrary
youtubeID to be added to our feeds.